New race: octopode
Based on http://crawl.chaosforge.org/Octopode to some extent.

-Can wear 8 rings.
--Extra rings fall off if polymorphed.
--Not a valid polymorph form, but it should work.
-Can be wizards, rogues, pirates, acus, or binders.
-Attribute caps:
 Str   Int Wis Dex Con Cha
 18/** 20  18  20  16  16
-Special spell: poison spray.
--Not very useful since they don't have a special way of learning it.
-Human HP and energy growth.
-Radius-2 low light vision.
-Medium sized.  Can wear barded helmets and soft hats, cloaks, dragon
 scales (not scale mail) and waistcloths, shields.
-Extra 1d8 tentacle wrap attack that can drown if over water.
-+level/10 bonus to natural DR and +level/2 bonus to natural AC, to
 compensate for being unable to wear most armour.
--Rings of protection can also compensate, but unlike in DCSS they
  need to be identified and charged first.
-Starts with stone resistance, poison resistance, stealth, swimming,
 magical breathing, waterproofing.
--Stone resistance because it would be unplayable otherwise (no
  gloves).
--Waterproof so swimming is usable.
-Immune to heart-piercing and bisection vorpality (doesn't have any
 special effect) and being ripped apart by Demogorgon (does double damage
 instead).
--Immunity to heart piercing vorpality could be explained by octopodes
  having 3 hearts, but the real reason for this is that they can't wear
  armour that would protect them.
-Weldproof because I didn't implement interactions of curses with
 being able to wear 8 rings.  Unlike the other intrinsics, this is
 only provided from form (you will lose it if you polymorph).
-Immune to glib because of their suckers (and because I didn't
 implement how being able to wear 8 rings interacts with it).
-Can't have rings put on by seduction because I didn't implement it.

Also:
-Parasitized androids are weldproof.
-Enlightenment message for white hlf waterproofing.

+++ include/obj.h
@@ -284,7 +284,7 @@ struct obj {
 	long age;		/* creation date */
 	/* in order to prevent alignment problems oextra should
 	   be (or follow) a long int */
-	long owornmask;
+	long long owornmask;
 	/* Weapons and artifacts */
 //	define W_ARM	    0x00000001L /* Body armor */
 //	define W_ARMC	    0x00000002L /* Cloak */
